refactor: RaceGame의 playRound에서 자동차 이동 책임을 Cars로 위임
